+++ Changelog.md
@@ -2,7 +2,7 @@
 
 ## Unreleased
 
-- Analysis: `--unpack-archives` now respects path filters ([#1745](https://github.com/fossas/fossa-cli/pull/1745))
+- Analysis: `--unpack-archives` now applies path filters to the archive's path in the scan ([#1745](https://github.com/fossas/fossa-cli/pull/1745))
 
 ## 3.17.17